1.008

+++ Src/Curve.h
@@ -29,7 +29,7 @@ typedef struct
 	int numOfTotalDots;								//理论计算总点数
 	int numOfPlotDots;								//已画点数量
 	int numOfDotsToPlot;							//将要画点数量
-	unsigned long time;								//时间
+	float time;										//时间
 	float* pDotXHead;								//点X坐标数组首地址   
 	float* pDotYHead;								//点Y坐标数组首地址  
 	float* pDotX;									//当前接收点X坐标地址
